solved86

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int count;
-	long long int num;
+	long int num;
 
 	num = 612852475143;
 	for (count = 2; num > 1; count++)
